+++ lgrez/__init__.py
@@ -8,9 +8,9 @@ See github.com/loic-simon/lg-rez for informations.
 __title__ = "lg-rez"
 __author__ = "Loïc Simon, Tom Lacoma"
 __license__ = "MIT"
-__copyright__ = ("Copyright 2020 - 2021 Loïc Simon & Tom Lacoma - "
+__copyright__ = ("Copyright 2020 - 2022 Loïc Simon & Tom Lacoma - "
                  "Club BD-Jeux × GRIs – ESPCI Paris - PSL")
-__version__ = "2.3.3"
+__version__ = "2.4.0"
 __all__ = ["LGBot"]