+++ lgrez/blocs/tools.py
@@ -150,7 +150,8 @@ def mention_MJ(arg):
         :class:`str`
     """
     member = arg.author if isinstance(arg, commands.Context) else arg
-    if (isinstance(member, discord.Member)
+    if (config.is_setup
+        and isinstance(member, discord.Member)
         and member.top_role >= config.Role.mj):
         # Pas un webhook et (au moins) MJ
         return f"@{config.Role.mj.name}"